Fix steps to show correct status
