+++ src/modules/InvitationForPunchOut/views/CreateAndEditIPO/EditIPO.tsx
@@ -1,4 +1,4 @@
-import { Attachment, CommPkgRow, GeneralInfoDetails, McScope, Participant, Person, RoleParticipant } from '../../types';
+import { Attachment, CommPkgRow, GeneralInfoDetails, McScope, Participant, Person, RoleParticipant, Step } from '../../types';
 import { CenterContainer, Container } from './CreateAndEditIPO.style';
 import { FunctionalRoleDto, ParticipantDto, PersonDto } from '../../http/InvitationForPunchOutApiClient';
 import React, { useCallback, useEffect, useState } from 'react';
@@ -28,6 +28,15 @@ const emptyGeneralInfo: GeneralInfoDetails = {
 };
 
 const EditIPO = (): JSX.Element => {
+
+    const initialSteps: Step[] = [
+        { title: 'General info', isCompleted: true },
+        { title: 'Scope', isCompleted: true },
+        { title: 'Participants', isCompleted: true },
+        { title: 'Upload attachments', isCompleted: true },
+        { title: 'Summary & update', isCompleted: true }
+    ];
+
     const params = useParams<{ ipoId: any }>();
     const [attachments, setAttachments] = useState<Attachment[]>([]);
     const [generalInfo, setGeneralInfo] = useState<GeneralInfoDetails>(emptyGeneralInfo);
@@ -48,6 +57,7 @@ const EditIPO = (): JSX.Element => {
     const { apiClient } = useInvitationForPunchOutContext();
     const { history } = useRouter();
     const { setDirtyStateFor, unsetDirtyStateFor } = useDirtyContext();
+    const [steps, setSteps] = useState<Step[]>(initialSteps);
 
     /**
      * Fetch and set available functional roles 
@@ -371,6 +381,8 @@ const EditIPO = (): JSX.Element => {
 
     return (<CreateAndEditIPO
         saveIpo={saveUpdatedIpo}
+        steps={steps}
+        setSteps={setSteps}
         generalInfo={generalInfo}
         setGeneralInfo={setGeneralInfo}
         selectedCommPkgScope={selectedCommPkgScope}
